Pet_Social/requirements.txt
Django==4.0.6
mysqlclient==2.1.0
djangorestframework==3.13.1
redis==4.3.4
Pillow==9.2.0
hiredis
uwsgi
